+++ components/additional_funcs.py
@@ -34,13 +34,13 @@ from discord import (
 from discord.abc import Messageable
 from discord.ext import commands
 from discord.ui import View, Select, Item, Button, button
-from discord.utils import get as utils_get, _get_mime_type_for_image
+from discord.utils import get as utils_get, _get_mime_type_for_image, MISSING
 from mcipc.query import Client as Client_q
 from mcipc.rcon import Client as Client_r, WrongPassword
 from psutil import process_iter, NoSuchProcess, disk_usage, Process, AccessDenied
 from requests import post as req_post, get as req_get, head as req_head, Timeout
 
-from components.decorators import MissingAdminPermissions
+from components.decorators import MissingAdminPermissions, is_admin, is_minecrafter
 from components.localization import get_translation, get_locales, get_current_locale, set_locale
 from components.rss_feed_handle import get_feed_webhook
 from components.watcher_handle import create_watcher, get_chat_webhook
@@ -49,6 +49,7 @@ from config.init_config import Config, BotVars, ServerProperties, OS, URL_REGEX
 if TYPE_CHECKING:
     from commands.poll import Poll
     from commands.minecraft_commands import MinecraftCommands
+    from commands.chat_commands import ChatCommands
 
 if Config.get_os() == OS.Windows:
     from os import startfile
@@ -595,6 +596,26 @@ async def get_role_string(bot: Union[commands.Bot, Client], id: int, mention: bo
     return role
 
 
+async def get_message_and_channel(bot: Union[commands.Bot, Client], message_id: int, channel_id: Optional[int] = None):
+    if channel_id is None:
+        for ch in bot.guilds[0].channels:
+            sub_chs = [ch]
+            if hasattr(ch, "threads"):
+                sub_chs.extend(ch.threads)
+            for sub_ch in sub_chs:
+                if isinstance(sub_ch, (TextChannel, VoiceChannel, Thread)):
+                    with suppress(NotFound, Forbidden, HTTPException):
+                        message = await sub_ch.fetch_message(message_id)
+                        return message, sub_ch
+        return None, None
+    else:
+        with suppress(NotFound, Forbidden, HTTPException, InvalidData):
+            channel = await bot.fetch_channel(channel_id)
+            message = await channel.fetch_message(message_id)
+            return message, channel
+        return None, None
+
+
 class BackupsThread(Thread):
     def __init__(self, bot):
         super().__init__()
@@ -883,11 +904,7 @@ def send_message_of_deleted_backup(bot: commands.Bot, reason: str, backup=None, 
             if member_name is not None:
                 member = member_name
             else:
-                member = bot.guilds[0].get_member(backup.initiator)
-                if member is not None:
-                    member = f"{member.display_name}#{member.discriminator}"
-                else:
-                    member = "invalid-user"
+                member = get_member_string(bot, backup.initiator)
             msg = get_translation("Deleted backup '{0}.zip' made by {1} because of {2}").format(backup.file_name,
                                                                                                 member, reason)
     else:
@@ -980,15 +997,18 @@ def get_human_readable_size(size: Union[int, float], stop_unit=None, round=False
         return f"{size:.2f} {get_translation(UNITS[-1])}"
 
 
-async def warn_about_auto_backups(ctx: commands.Context, bot: commands.Bot):
+async def warn_about_auto_backups(ctx: Union[commands.Context, Interaction], bot: commands.Bot, is_reaction=False):
     if Config.get_backups_settings().automatic_backup:
         if len(Config.get_server_config().backups) > 0 \
                 and handle_backups_limit_and_size(bot) is not None \
                 and all([b.initiator for b in Config.get_server_config().backups]):
-            await ctx.send(get_translation("Bot has backups only from members for '{0}' server, "
-                                           "so keep in mind, that bot will delete oldest backup "
-                                           "on next auto backup!")
-                           .format(Config.get_selected_server_from_list().server_name))
+            await send_msg(
+                ctx,
+                get_translation("Bot has backups only from members for '{0}' server, "
+                                "so keep in mind, that bot will delete oldest backup "
+                                "on next auto backup!").format(Config.get_selected_server_from_list().server_name),
+                is_reaction=is_reaction
+            )
 
 
 def get_half_members_count_with_role(channel: TextChannel, role: int):
@@ -1466,7 +1486,7 @@ async def bot_backup(
                            (backup.reason if backup.reason else get_translation("Not stated"))
             bot_message += "\n" + get_translation("Initiator: ") + await get_member_string(bot, backup.initiator)
         if backup.restored_from:
-            bot_message += "\n\t" + get_translation("The world of the server was restored from this backup")
+            bot_message += "\n" + get_translation("The world of the server was restored from this backup")
     await send_msg(ctx, add_quotes(bot_message), is_reaction)
 
 
@@ -1865,6 +1885,7 @@ class TemplateSelectView(View):
             namespace: str = "template_select_view",
             name: str = "TemplateSelectView",
             ctx: Union[commands.Context, TextChannel, VoiceChannel, ChannelThread, GroupChannel, None] = None,
+            raw_content: Optional[str] = None,
             message_id: Optional[int] = None,
             channel_id: Optional[int] = None,
             message: Optional[Message] = None,
@@ -1877,10 +1898,11 @@ class TemplateSelectView(View):
         super().__init__(timeout=timeout)
         self.namespace = namespace
         self.name = name
-        self.ctx=ctx
+        self.ctx = ctx
         self.message_id = message_id
         self.channel_id = channel_id
         self.message = message
+        self.raw_content = raw_content
         self.select_page: int = -1
         self.options_list_raw = options_raw
         self.options_list_length = len(options_raw)
@@ -1903,7 +1925,7 @@ class TemplateSelectView(View):
         self.v_start = Button(
             style=ButtonStyle.secondary,
             custom_id=f"{namespace}:select_start",
-            emoji="⏪",
+            emoji="⏮",
             row=start_row + 1
         )
         self.v_start.callback = self.v_start_callback
@@ -1912,7 +1934,7 @@ class TemplateSelectView(View):
         self.v_left = Button(
             style=ButtonStyle.secondary,
             custom_id=f"{namespace}:select_left",
-            emoji="⬅",
+            emoji="◀",
             row=start_row + 1
         )
         self.v_left.callback = self.v_left_callback
@@ -1930,7 +1952,7 @@ class TemplateSelectView(View):
         self.v_right = Button(
             style=ButtonStyle.secondary,
             custom_id=f"{namespace}:select_right",
-            emoji="➡",
+            emoji="▶",
             row=start_row + 1
         )
         self.v_right.callback = self.v_right_callback
@@ -1939,13 +1961,18 @@ class TemplateSelectView(View):
         self.v_end = Button(
             style=ButtonStyle.secondary,
             custom_id=f"{namespace}:select_end",
-            emoji="⏩",
+            emoji="⏭",
             row=start_row + 1
         )
         self.v_end.callback = self.v_end_callback
         self.add_item(self.v_end)
 
-    async def update_view(self, send: bool = True):
+    async def update_view(
+            self,
+            send: bool = True,
+            update_content: bool = False,
+            check_if_content_is_different: bool = False
+    ):
         await self.update_view_components()
         await self.update_select_options()
         if self.options_list_length > DISCORD_SELECT_OPTIONS_MAX_LENGTH:
@@ -1978,21 +2005,21 @@ class TemplateSelectView(View):
             return
 
         if self.message is None and self.message_id is not None:
-            if self.channel_id is None:
-                for ch in BotVars.bot_for_webhooks.guilds[0].channels:
-                    sub_chs = [ch]
-                    if hasattr(ch, "threads"):
-                        sub_chs.extend(ch.threads)
-                    for sub_ch in sub_chs:
-                        if isinstance(sub_ch, (TextChannel, VoiceChannel, Thread)):
-                            with suppress(NotFound, Forbidden, HTTPException):
-                                self.message = await sub_ch.fetch_message(self.message_id)
-                                self.channel_id = sub_ch.id
-            else:
-                channel = await BotVars.bot_for_webhooks.fetch_channel(self.channel_id)
-                self.message = await channel.fetch_message(self.message_id)
+            self.message, channel = await get_message_and_channel(
+                BotVars.bot_for_webhooks,
+                self.message_id,
+                self.channel_id
+            )
+            if channel is not None:
+                self.channel_id = channel.id
         if self.message is not None:
-            await self.message.edit(view=self)
+            if check_if_content_is_different and self.message.clean_content == get_translation(self.raw_content):
+                update_content = False
+            self.message = await self.message.edit(
+                view=self,
+                content=(get_translation(self.raw_content)
+                         if self.raw_content is not None and update_content else MISSING)
+            )
         await self.do_after_sending_message()
 
     async def update_select_options(self, page: Optional[int] = None):
@@ -2120,13 +2147,12 @@ class SelectView(TemplateSelectView):
 
     async def update_select_options(self, page: Optional[int] = None):
         if page is None:
-            self.v_select.options = [
-                await self.make_select_option(i, self.bot) for i in range(*self.get_indexes(self.pivot_index))
-            ]
+            indexes = self.get_indexes(self.pivot_index)
         else:
-            self.v_select.options = [
-                await self.make_select_option(i, self.bot) for i in range(*self.set_by_page(page))
-            ]
+            indexes = self.set_by_page(page)
+        self.v_select.options = [
+            await self.make_select_option(i, self.bot) for i in range(*indexes)
+        ]
 
 
 class SelectChoice(Enum):
@@ -2232,6 +2258,8 @@ class MenuServerView(TemplateSelectView):
     def __init__(self, commands_cog: 'MinecraftCommands'):
         super().__init__(
             options_raw=Config.get_settings().servers_list,
+            raw_content="List of commands for interacting with Minecraft server via buttons"
+                        " and dropdown for selecting server",
             namespace="menu_server_view",
             name="MenuServerView",
             message_id=Config.get_menu_settings().server_menu_message_id,
@@ -2243,23 +2271,17 @@ class MenuServerView(TemplateSelectView):
 
     async def update_select_options(self, page: Optional[int] = None):
         if page is None:
-            self.v_select.options = [
-                SelectOption(
-                    label=shorten_string(Config.get_settings().servers_list[i].server_name,
-                                         DISCORD_SELECT_FIELD_MAX_LENGTH),
-                    value=str(i),
-                    default=i + 1 == Config.get_settings().selected_server_number
-                ) for i in range(*self.get_indexes(Config.get_settings().selected_server_number))
-            ]
+            indexes = self.get_indexes(Config.get_settings().selected_server_number)
         else:
-            self.v_select.options = [
-                SelectOption(
-                    label=shorten_string(Config.get_settings().servers_list[i].server_name,
-                                         DISCORD_SELECT_FIELD_MAX_LENGTH),
-                    value=str(i),
-                    default=i + 1 == Config.get_settings().selected_server_number
-                ) for i in range(*self.set_by_page(page))
-            ]
+            indexes = self.set_by_page(page)
+        self.v_select.options = [
+            SelectOption(
+                label=shorten_string(Config.get_settings().servers_list[i].server_name,
+                                     DISCORD_SELECT_FIELD_MAX_LENGTH),
+                value=str(i),
+                default=i + 1 == Config.get_settings().selected_server_number
+            ) for i in range(*indexes)
+        ]
 
     async def do_after_sending_message(self):
         if self.channel_id is not None and Config.get_menu_settings().server_menu_channel_id is None:
@@ -2267,21 +2289,9 @@ class MenuServerView(TemplateSelectView):
             Config.save_config()
 
     async def interaction_check_select(self, interaction: Interaction, /) -> bool:
-        if Config.get_settings().bot_settings.managing_commands_role_id is None or \
-                Config.get_settings().bot_settings.managing_commands_role_id \
-                in (e.id for e in interaction.user.roles):
-            return True
-        else:
-            await send_error_on_interaction(
-                self.name,
-                interaction,
-                ctx=None,
-                error=commands.MissingRole(Config.get_settings().bot_settings.managing_commands_role_id),
-                is_reaction=True
-            )
-            return False
+        return is_minecrafter(interaction)
 
-    @button(label="status", style=ButtonStyle.secondary, custom_id="menu_server_view:status", emoji="🗨", row=0)
+    @button(label="status", style=ButtonStyle.secondary, custom_id="menu_server_view:status", emoji="⚠", row=0)
     async def c_status(self, interaction: Interaction, button: Button):
         BotVars.react_auth = interaction.user
         await bot_status(interaction, interaction.client, is_reaction=True)
@@ -2301,10 +2311,10 @@ class MenuServerView(TemplateSelectView):
         self.commands_cog.checkups_task.restart()
         await send_interaction(interaction, get_translation("Updated bot status!"), is_reaction=True)
 
-    @button(label="start", style=ButtonStyle.secondary, custom_id="menu_server_view:start", emoji="♿", row=1)
+    @button(label="start", style=ButtonStyle.secondary, custom_id="menu_server_view:start", emoji="⏯", row=1)
     async def c_start(self, interaction: Interaction, button: Button):
         BotVars.react_auth = interaction.user
-        if self.interaction_check_select(interaction):
+        if await self.interaction_check_select(interaction):
             await bot_start(
                 interaction,
                 interaction.client,
@@ -2343,6 +2353,252 @@ class MenuServerView(TemplateSelectView):
             await self.update_view()
 
 
+async def on_language_select_callback(
+        interaction: Optional[Interaction],
+        set_language: Optional[str],
+        ctx: Optional[commands.Context] = None,
+        is_reaction: bool = False
+):
+    new_language = interaction.data.get("values", [None])[0] if interaction is not None else set_language
+
+    if not set_locale(new_language):
+        msg = add_quotes(get_translation("Bot doesn't have this language!\n"
+                                         "Check list of available languages via {0}language")
+                         .format(Config.get_settings().bot_settings.prefix))
+        await send_interaction(interaction, msg, ctx=ctx, is_reaction=is_reaction)
+        return SelectChoice.DO_NOTHING
+    else:
+        Config.get_settings().bot_settings.language = new_language.lower()
+        Config.save_config()
+        await send_interaction(
+            interaction,
+            add_quotes(get_translation("Language switched successfully!")),
+            ctx=ctx,
+            is_reaction=is_reaction
+        )
+        return SelectChoice.STOP_VIEW
+
+
+class MenuBotView(TemplateSelectView):
+    def __init__(self, bot: commands.Bot, commands_cog: 'MinecraftCommands'):
+        super().__init__(
+            options_raw=get_locales(),
+            raw_content="List of bot features for interaction via buttons"
+                        " and dropdown for selecting bot language",
+            namespace="menu_bot_view",
+            name="MenuBotView",
+            message_id=Config.get_menu_settings().bot_menu_message_id,
+            channel_id=Config.get_menu_settings().bot_menu_channel_id,
+            start_row=3,
+            is_reaction=True
+        )
+        self.bot = bot
+        self.commands_cog = commands_cog
+        self.chat_cog: Optional['ChatCommands'] = None
+
+    async def update_select_options(self, page: Optional[int] = None):
+        pivot = [i for i in range(self.options_list_length) if get_current_locale() == self.options_list_raw[i]][0]
+        if page is None:
+            indexes = self.get_indexes(pivot)
+        else:
+            indexes = self.set_by_page(page)
+        self.v_select.options = [
+            SelectOption(
+                label=shorten_string(self.options_list_raw[i].capitalize(), DISCORD_SELECT_FIELD_MAX_LENGTH),
+                value=shorten_string(self.options_list_raw[i], DISCORD_SELECT_FIELD_MAX_LENGTH),
+                description=shorten_string(get_translation(self.options_list_raw[i]),
+                                           DISCORD_SELECT_FIELD_MAX_LENGTH),
+                default=pivot == i,
+                emoji="🌐" if pivot == i else None
+            ) for i in range(*indexes)
+        ]
+
+    async def update_view_components(self):
+        if self.chat_cog is None:
+            self.chat_cog = self.bot.get_cog("ChatCommands")
+            if self.chat_cog is None:
+                raise RuntimeError("Cog 'ChatCommands' not found!")
+
+        self.b_auth.style = ButtonStyle.green if Config.get_secure_auth().enable_secure_auth else ButtonStyle.red
+        self.b_auth.emoji = "🔒" if Config.get_secure_auth().enable_secure_auth else "🔑"
+        self.b_auth.label = get_translation("Authorization")
+
+        self.b_a_backup.style = ButtonStyle.green if Config.get_backups_settings().automatic_backup else ButtonStyle.red
+        self.b_a_backup.label = get_translation("Auto backup")
+
+        self.b_whitelist.style = ButtonStyle.green if ServerProperties().white_list else ButtonStyle.red
+        self.b_whitelist.label = get_translation("White list")
+
+        self.b_forceload.style = ButtonStyle.green if Config.get_settings().bot_settings.forceload else ButtonStyle.red
+        self.b_forceload.emoji = "♾" if Config.get_settings().bot_settings.forceload else "🇽"
+        self.b_forceload.label = get_translation("Forceload")
+
+        self.b_a_shutdown.style = ButtonStyle.green \
+            if Config.get_settings().bot_settings.auto_shutdown else ButtonStyle.red
+        self.b_a_shutdown.emoji = "🌜" if Config.get_settings().bot_settings.auto_shutdown else "🌕"
+        self.b_a_shutdown.label = get_translation("Auto shutdown")
+
+        self.b_chat.style = ButtonStyle.green \
+            if Config.get_cross_platform_chat_settings().enable_cross_platform_chat else ButtonStyle.red
+        self.b_chat.label = get_translation("C/p chat")
+
+        self.b_c_p_images.style = ButtonStyle.green \
+            if Config.get_cross_platform_chat_settings().image_preview.enable_images_preview else ButtonStyle.red
+        self.b_c_p_images.label = get_translation("Image preview")
+
+        self.b_rss_news.style = ButtonStyle.green if Config.get_rss_feed_settings().enable_rss_feed else ButtonStyle.red
+        self.b_rss_news.emoji = "🔔" if Config.get_rss_feed_settings().enable_rss_feed else "🔕"
+        self.b_rss_news.label = get_translation("RSS feed")
+
+    async def do_after_sending_message(self):
+        if self.channel_id is not None and Config.get_menu_settings().bot_menu_channel_id is None:
+            Config.get_menu_settings().bot_menu_channel_id = self.channel_id
+            Config.save_config()
+
+    @button(custom_id="menu_bot_view:authorize", row=0)
+    async def b_auth(self, interaction: Interaction, button: Button):
+        if is_admin(interaction):
+            Config.get_secure_auth().enable_secure_auth = not Config.get_secure_auth().enable_secure_auth
+            Config.save_config()
+            button.emoji = "🔑" if button.style == ButtonStyle.green else "🔒"
+            button.style = ButtonStyle.red if button.style == ButtonStyle.green else ButtonStyle.green
+            await edit_interaction(interaction, self, self.message_id)
+            if Config.get_secure_auth().enable_secure_auth:
+                with suppress(ConnectionError, socket.error):
+                    if BotVars.watcher_of_log_file is None:
+                        BotVars.watcher_of_log_file = create_watcher(BotVars.watcher_of_log_file, get_server_version())
+                    BotVars.watcher_of_log_file.start()
+                msg = add_quotes(get_translation("Secure authorization on"))
+            else:
+                if not Config.get_cross_platform_chat_settings().enable_cross_platform_chat and \
+                        BotVars.watcher_of_log_file is not None:
+                    BotVars.watcher_of_log_file.stop()
+                msg = add_quotes(get_translation("Secure authorization off"))
+            await send_interaction(interaction, msg, is_reaction=True)
+
+    @button(custom_id="menu_bot_view:auto_backup", emoji="💾", row=0)
+    async def b_a_backup(self, interaction: Interaction, button: Button):
+        if is_minecrafter(interaction):
+            Config.get_backups_settings().automatic_backup = not Config.get_backups_settings().automatic_backup
+            Config.save_config()
+            button.style = ButtonStyle.red if button.style == ButtonStyle.green else ButtonStyle.green
+            await edit_interaction(interaction, self, self.message_id)
+            if Config.get_backups_settings().automatic_backup:
+                await warn_about_auto_backups(interaction, self.bot, is_reaction=True)
+                msg = add_quotes(get_translation("Automatic backups enabled"))
+            else:
+                msg = add_quotes(get_translation("Automatic backups disabled"))
+            await send_interaction(interaction, msg, is_reaction=True)
+
+    @button(custom_id="menu_bot_view:whitelist", emoji="🗒️", row=0)
+    async def b_whitelist(self, interaction: Interaction, button: Button):
+        if is_minecrafter(interaction):
+            async with handle_rcon_error(None, interaction, is_reaction=True):
+                msg = None
+                with connect_rcon() as cl_r:
+                    if ServerProperties().white_list:
+                        cl_r.run("whitelist off")
+                        msg = add_quotes(get_translation("The server is allowed to let any players regardless "
+                                                         "of the list of allowed nicknames"))
+                    else:
+                        cl_r.run("whitelist on")
+                        msg = add_quotes(get_translation("The server is forbidden to let players not "
+                                                         "from the list of allowed nicknames"))
+
+                button.style = ButtonStyle.red if button.style == ButtonStyle.green else ButtonStyle.green
+                await edit_interaction(interaction, self, self.message_id)
+                if msg is not None:
+                    await send_interaction(interaction, msg, is_reaction=True)
+
+    @button(custom_id="menu_bot_view:forceload", row=1)
+    async def b_forceload(self, interaction: Interaction, button: Button):
+        if is_minecrafter(interaction):
+            Config.get_settings().bot_settings.forceload = not Config.get_settings().bot_settings.forceload
+            Config.save_config()
+            button.emoji = "🇽" if button.style == ButtonStyle.green else "♾"
+            button.style = ButtonStyle.red if button.style == ButtonStyle.green else ButtonStyle.green
+            await edit_interaction(interaction, self, self.message_id)
+            await send_interaction(interaction, add_quotes(bot_forceload_info()), is_reaction=True)
+
+    @button(custom_id="menu_bot_view:auto_shutdown", row=1)
+    async def b_a_shutdown(self, interaction: Interaction, button: Button):
+        if is_minecrafter(interaction):
+            Config.get_settings().bot_settings.auto_shutdown = not Config.get_settings().bot_settings.auto_shutdown
+            Config.save_config()
+            button.emoji = "🌕" if button.style == ButtonStyle.green else "🌜"
+            button.style = ButtonStyle.red if button.style == ButtonStyle.green else ButtonStyle.green
+            await edit_interaction(interaction, self, self.message_id)
+            await send_interaction(
+                interaction,
+                add_quotes(bot_shutdown_info(with_timeout=Config.get_settings().bot_settings.auto_shutdown)),
+                is_reaction=True
+            )
+
+    @button(custom_id="menu_bot_view:chat", emoji="💬", row=2)
+    async def b_chat(self, interaction: Interaction, button: Button):
+        if is_admin(interaction):
+            Config.get_cross_platform_chat_settings().enable_cross_platform_chat = \
+                not Config.get_cross_platform_chat_settings().enable_cross_platform_chat
+            Config.save_config()
+            button.style = ButtonStyle.red if button.style == ButtonStyle.green else ButtonStyle.green
+            await edit_interaction(interaction, self, self.message_id)
+            if Config.get_cross_platform_chat_settings().enable_cross_platform_chat:
+                BotVars.webhook_chat = None
+                await create_webhooks(self.bot)
+                with suppress(ConnectionError, socket.error):
+                    if BotVars.watcher_of_log_file is None:
+                        BotVars.watcher_of_log_file = create_watcher(BotVars.watcher_of_log_file, get_server_version())
+                    BotVars.watcher_of_log_file.start()
+                msg = get_translation("Cross-platform chat enabled") + "!"
+            else:
+                if not Config.get_secure_auth().enable_secure_auth and BotVars.watcher_of_log_file is not None:
+                    BotVars.watcher_of_log_file.stop()
+                msg = get_translation("Cross-platform chat disabled") + "!"
+            await send_interaction(interaction, msg, is_reaction=True)
+
+    @button(custom_id="menu_bot_view:chat_preview_images", emoji="🖼", row=2)
+    async def b_c_p_images(self, interaction: Interaction, button: Button):
+        if is_minecrafter(interaction):
+            Config.get_cross_platform_chat_settings().image_preview.enable_images_preview = \
+                not Config.get_cross_platform_chat_settings().image_preview.enable_images_preview
+            Config.save_config()
+            button.style = ButtonStyle.red if button.style == ButtonStyle.green else ButtonStyle.green
+            await edit_interaction(interaction, self, self.message_id)
+            if Config.get_cross_platform_chat_settings().image_preview.enable_images_preview:
+                msg = get_translation("Image preview enabled") + "!"
+            else:
+                msg = get_translation("Image preview disabled") + "!"
+            await send_interaction(interaction, msg, is_reaction=True)
+
+    @button(custom_id="menu_bot_view:rss_news", row=2)
+    async def b_rss_news(self, interaction: Interaction, button: Button):
+        if is_admin(interaction):
+            Config.get_rss_feed_settings().enable_rss_feed = not Config.get_rss_feed_settings().enable_rss_feed
+            Config.save_config()
+            button.emoji = "🔕" if button.style == ButtonStyle.green else "🔔"
+            button.style = ButtonStyle.red if button.style == ButtonStyle.green else ButtonStyle.green
+            await edit_interaction(interaction, self, self.message_id)
+            if Config.get_rss_feed_settings().enable_rss_feed:
+                BotVars.webhook_rss = None
+                await create_webhooks(self.bot)
+                if self.chat_cog.rss_feed_task.is_running():
+                    self.chat_cog.rss_feed_task.restart()
+                else:
+                    self.chat_cog.rss_feed_task.start()
+                msg = get_translation("RSS enabled") + "!"
+            else:
+                if self.chat_cog.rss_feed_task.is_running():
+                    self.chat_cog.rss_feed_task.stop()
+                msg = get_translation("RSS disabled") + "!"
+            await send_interaction(interaction, msg, is_reaction=True)
+
+    async def v_select_callback(self, interaction: Interaction):
+        await on_language_select_callback(interaction, None, is_reaction=True)
+        await self.update_view(update_content=True)
+        if self.commands_cog.menu_server_view is not None:
+            await self.commands_cog.menu_server_view.update_view(update_content=True)
+
+
 @contextmanager
 def connect_rcon(timeout=1):
     try:
@@ -2366,15 +2622,24 @@ def connect_query():
 
 
 @asynccontextmanager
-async def handle_rcon_error(ctx: commands.Context, interaction: Interaction = None):
+async def handle_rcon_error(ctx: Optional[commands.Context], interaction: Interaction = None, is_reaction=False):
     try:
         yield
     except (ConnectionError, socket.error):
         if BotVars.is_server_on:
-            await send_interaction(interaction,
-                                   add_quotes(get_translation("Couldn't connect to server, try again(")), ctx=ctx)
+            await send_interaction(
+                interaction,
+                add_quotes(get_translation("Couldn't connect to server, try again(")),
+                ctx=ctx,
+                is_reaction=is_reaction
+            )
         else:
-            await send_interaction(interaction, add_quotes(get_translation("server offline").capitalize()), ctx=ctx)
+            await send_interaction(
+                interaction,
+                add_quotes(get_translation("server offline").capitalize()),
+                ctx=ctx,
+                is_reaction=is_reaction
+            )
 
 
 class HelpCommandArgument(commands.CheckFailure):